Explicitly wait on existing spans to finish before OTEL shutdown
Currently, none of our OTEL spans outlive the overall processing
of the original HTTP request. However, once we introduce spans
for rate-limiting and first-of-n, we'll start exporting spans for
background logic (e.g. `return_tickets`).

To make sure that we don't skip exporting any spans, we now attach
a `TaskTrackerToken` to the OpenTelmemetry `Context`, and wait
for the parent `TaskTracker` to finish before shutting down any
OpenTelemetry trackers. The `tracing-opentelemetry` crate will
propagate the `Context` to child spans, ensuring that the
`TaskTrackerToken` is only dropped after all of the descendant
spans are cloesd (e.g. for instrumented `tokio::spawn` tasks).