Add 'deferred_tasks' to AppStateData and use it for rate-limiting (#3971)
* Add 'deferred_tasks' to AppStateData and use it for rate-limiting

This is a new `TaskTracker` that we wait for when the gateway
shuts down. This ensures that we return rate-limiting tickets for all processed
requests.

Testing this is tricky at the moment (since we don't have a good way
of detecting whether or not we wrote to Postgres), but it will be
easier to test this behavior once we start emitting rate-limit spans

* Skip wait when tracker is empty