Lint 'tokio::task::spawn' and add TODO comments on existing code (#3988)
Whenever we call 'tokio::task::spawn', we need to make sure that it's
okay for the gateway to shut down without waiting on the task to finish.
This lint cannot catch all cases (we might call a wrapper from another
crate), but it's a significant improvement.

I've suppressed the lint at all existing callsites, leaving a TODO
at each callsite. Over time, we should update each comment to explain
why it's safe, or use a `TaskTracker` that we wait on during gateway
shutdown.

All new usages of 'tokio::task::spawn' should come with an explanation
of why we don't need to wait on the task during shutdown.