Add Postgres support to 'tensorzero-auth' API keys
We now expose functions for creating, checking, disabling, and listing
API keys in Postgres.

This is not yet hooked up to TensorZero itself, but we're using alpha
`sqlx` so that we can override the '_sqlx_migrations' table name.
This allows us to use an independent sqlx 'Migrator' instance for
'tensorzero-auth', without interfering with the existing postgres
migrations in `tensorzero-core`