Log a warning when a client closes the request early
Previously, we wouldn't log anything if Axum dropped our
future before we sent back a response (due to the client closing
the underlying connection).

We now log a warning if the entire response body is not sent to
the client before it gets dropped. This correctly handles streaming
responses, where we want to warn if the client closes the connection
before we've sent all of the SSE lines