Suppress all otel spans that do not descend from an HTTP span
This ensures that if a non-instrmented route (e.g. a ui route)
calls into code that emits otel spans, those spans are *not* emitted.

This gives us tighter control over exactly when we export things to
otel. In particular, we'll be able to add otel spans for authorization
middleware, without those spans getting exported if the authorization
middleware is applied to a non-otel route (e.g. the ui)