Add 'tensorzero-otlp-traces-extra-attribute-' header support
This header attaches an extra OpenTelemetry span attribute
to *every* span created as part of the HTTP request processing.
This is similar to 'tensorzero-otlp-traces-extra-resource-',
but this explicitly does not use the existing OpenTelemetry 'resource'
concept. Unfortunately, some providers like Arize don't display custom
resources, so we need to explicitly attach span attributes to get
custom information to show up in the UI.

The header value must be a valid JSON value, which we then convert
to an OpenTelemetry value (rejecting things like 'null' and objects).
In particular, this means that a string value must be quoted in the
header (unlike 'tensorzero-otlp-traces-extra-resource-'):

```
tensorzero-otlp-traces-extra-attribute-my-attr: "My quoted string"
```

This allows us to support setting boolean attributes, and is
forwards-compatible with supporting OpenTelemetry homoegeneous
arrays.