Fix merge_group detection in our workflows (#4155)
* Fix merge_group detection in our workflows

When we invoke a 'workflow_call' workflow with a 'uses' step,
the event will not be 'merge_group' in the callee workflow.
Instead, we need to pass in an explicit input parmaeter
telling the callee workflow if it's running in merge_group mode or not.

This was causing us to not run the ui clickhouse tests against multiple
clickhouse version in the merge group

* Fix input args