Add 'unbounded_recursion_wrapper' helper to prevent stack overflow
We use this inside any `async fn`s where we might trigger unbounded
recursion (currently, just `best_of_n` and `mixture_of_n`).
This has the side effect of fixing a stack overflow in our test
for 'run_evaluations_best_of_3' and 'run_evaluations_mixture_of_3'.
It's not clear why the stack overflow was inconsistent, but it's likely
due to the exact amount of stack space that rustc/llvm end up choosing
(which was running up against the stack limit in debug mode)

This unfortunately requires a `Send + 'static` bound, and some
refactoring to work around an annoying compiler error.