Add TensorZero API key auth middleware
We have a new `gateway.auth.enabled` config setting.
When set to 'true', all routes except for /health and /status
now require an 'Authorization: Bearer <API_KEY>' header.
This includes UI routes, so the ui will need changes to
work when auth is enabled.

The auth middleware manually excludes the /health and /status
routes, and otherwise checks that the API key exists in the databsae
and is not disabled (using the existing `tensorzero_auth` API)