Add explicit `Config::new_empty` method
We now use this whenever we need to construct an empty config
in production code. The `Default` impl is now gated behind
`#[cfg(test)]` - it cannot run all of the necessary logic,
since `default` is not `async`. However, all of our existing tests
don't actually need a valid config, so it's fine to use it for
convenience in tests.

This is preparation for adding built-in functions, which will result
in an empty config file diverging from `Config::default()` (since
built-in functions will only be inserted during the normal config
loading logic)