Use `DelayedError` when constructing credential errors
This allows us to avoid logging an error if we end up using
a fallback credential. Inside of each provider, we covert
the `DelayedError` to an `Error` using `DelayedError.log`
at the point where we actually need the credentials
(e.g. before performing an inference).

I've also turned off the 'redundant_closure_for_method_calls' lint
globally, as it produces lots of noise, and doesn't seem very helpful
in general.