Don't drop the 'uint_to_uuid' function in rollback instructions (#4382)
ClickHouse user-defined functions are created globally, and are
not scoped to any database. As a result, trying to drop them in
the rollback instructions will interfere with any other migration
running currently on the same instance. In particular, this makes
our migration-manager tests flaky.

We no longer drop this function - it's fine to leave a single
function hanging around if the user does a manual rollback.