Reject request if creating custom OTEL tracer fails (#4476)
Using a custom tracer can be required for traces to be correctly
exported (e.g. with Arize), so an error creating a custom tracer
might only ever show up in the logs. Let's reject the request when
this happens, instead of continuing on and potentially failing to
export the inference trace to OTEL.