Add `Error.underlying_status_code`
This gives us the last error status in the case of a 'nested'
error. For example, if all of the models fail with an HTTP
error, then we'll return the status code of the last model
request.

We don't use this yet, but we'll eventually expose those
in some way in our error messages (either directly setting the
HTTP response status, or exposing it as a field)