Enable clippy::large_futures lint and box some futures
We're seeing stack overflow errors in the tensorzreo_relay
PR, which these additional Box's fix. I'm not certain if
this is only due to the additional stack usage in debug builds,
or if there's a risk of hitting this in production. To be on the
same side, let's enable the lint