Re-create 'ClickHouseConnectionInfo' in ClientBuilderMode::FromComponents
Using the same 'ClickHouseConnectionInfo' in multiple distinct embedded
clients is incorrect, as it interacts badly with things like
batch-writes and block-on-shutdown. Instead, we construct a fresh
'ClickHouseConnectionInfo' that re-uses the same database settings.
This is the equivalent of doing 'ClickHouseConnectionInfo::new'
with the same database url and batch config, but allows us to avoid
needing to carry those fields around separately.

This will allow us to use 'ClientBuilderMode::FromComponents' from
Python without causing deadlocks when a client gets dropped with
the GIL held