Compute stream TTFT based on first (possibly empty) chunk
By prefilling with 'assistant' messages, it's possible to make
some models produce a completely empty stream response
(all chunks have empty content arrays). We now treat the arrival
of this chunk as the first 'token' (the invisible stop-sequence token).

As a result, we now create database entries for completely empty stream
responses (matching what we do in the non-streaming case)