Bump tensorzero-core to edition 2024
This let us use 'if-let-chains'
(`if let Some(a) = foo && a.some_variable = "something"`)

Unfortunately, the diff is quite large, since it looks like
'cargo fmt' does different things on the new edition.

The only nontrivial change is the use of the (test-only)
`env` dependency. The 'std::env::{set_var, remove_var}'
functions are now `unsafe` in edition 2024. In order
to keep using `forbid(unsafe_code)`, I've added a dependency
on the 'env' crate, which provides safe replacements.