Make provider-proxy test_read_old_write_new deterministic
Previously, we weren't waiting on the file to get updated on disk
before sending a new request. As a result, the new file could get
written out after the new provider-proxy server started, resulting in
the file mtime being newer than than the start time.

We now wait on the file mtime to change (and for the target server to
shut down) before launching a new proxy-proxy server