Propagate OpenTelemetry context on all outgoing HTTP requests
In particular, this sets a 'traceparent' header corresponding
to the current active OpenTelemetry span. This allows the model
host to publish OpenTelemetry spans rooted at the existing
'model_provider_inference' span.

This will be particularly uesful for the relay mode, so that we
can obtain properly-connected traces if all of the gateways
export to the same OpenTelemetry endpoint

For now, I'm testing this by adding a loopback openai-compatible
model definition in our toml file, which points back at the gateway's
openai-compatible endpoint with 'dummy::good'