Add 'sanitize_traceparent' option to provider-proxy
We now set the 'traceparent' header on all outgoing HTTP requests
when OTLP is enabled.
However, the value of this header is random, which breaks provider-proxy
caching.

The new 'sanitize_traceparent' works like the existing 'sanitize_*'
options - we replace the header with a fixed value when computing
the cache key in provider-proxy