Add 'tensorzero_overhead' histogram metric
This metric tracks the 'overhead' of TensorZero on each HTTP request.
We calculate this by computing the total request processing time
(from span creation to the time when the response body is fully sent to
the client), and subtract off the time intervals where an
external HTTP request was active.