Only write streaming provider-proxy cache body on finish
We were previously writing the current body when the stream was
dropped, even if it never finished (e.g. due to a timeout).
As a result, we could end up writing an invalid body to disk
due to an *earlier* request from the e2e tests (if the e2e tests
hit an internal timeout and dropped the stream), overwriting
the good cache line from a later successful test run