Add response wrapper type to `TensorzeroHttpClient`
Previously, calling the `send` method would return a
`reqwest::Response` and drop our `LimitedClientTicket`.
This will incorrectly signal that we now have additional
space in the pool (even though the HTTP connection is still
needed to read the response body). It also will interfere with
storing a `Span` to track the total HTTP duration (for the upcoming
'tensorzero_overhead' metric)

I've introduced a new wrapper type. It's similar to
`TensorzeroRequestBuilder` - we hold on to our `LimitedClientTicket`,
and expose accessor methods that forward to the underlying type.